requirements.txt
Flask>=2.0
segno>=1.4
Pillow>=9.0
